feat: 카테고리 및 워크스페이스 관련 경로 수정
