+++ src/main/java/courseitda/workspace/application/dto/request/DeleteCategoryCommand.java
@@ -4,7 +4,6 @@ import courseitda.auth.domain.MemberAuthInfo;
 
 public record DeleteCategoryCommand(
         MemberAuthInfo memberAuthInfo,
-        String workspaceIdentifier,
         Long categoryId
 ) {
 }
