+++ src/main/java/courseitda/workspace/application/dto/request/FindCategoryCommand.java
@@ -4,7 +4,6 @@ import courseitda.auth.domain.MemberAuthInfo;
 
 public record FindCategoryCommand(
         MemberAuthInfo memberAuthInfo,
-        String workspaceIdentifier,
         Long categoryId
 ) {
 }
