+++ src/main/java/courseitda/workspace/application/dto/request/UpdateCategoryCommand.java
@@ -4,7 +4,6 @@ import courseitda.auth.domain.MemberAuthInfo;
 
 public record UpdateCategoryCommand(
         MemberAuthInfo memberAuthInfo,
-        String workspaceIdentifier,
         Long categoryId,
         String name,
         String color
