+++ src/main/java/courseitda/workspace/ui/CategoryController.java
@@ -5,25 +5,17 @@ import courseitda.auth.domain.MemberAuthInfo;
 import courseitda.auth.domain.RequiresRole;
 import courseitda.workspace.application.CategoryService;
 import courseitda.workspace.application.dto.request.DeleteCategoryCommand;
-import courseitda.workspace.application.dto.request.FindAllCategoriesCommand;
 import courseitda.workspace.application.dto.request.FindCategoryCommand;
-import courseitda.workspace.ui.dto.request.CategoryCreateRequest;
-import courseitda.workspace.ui.dto.request.CategoryReorderRequest;
 import courseitda.workspace.ui.dto.request.CategoryUpdateRequest;
-import courseitda.workspace.ui.dto.response.CategoriesResponse;
-import courseitda.workspace.ui.dto.response.CategoryCreateResponse;
-import courseitda.workspace.ui.dto.response.CategoryReorderResponse;
 import courseitda.workspace.ui.dto.response.CategoryResponse;
 import courseitda.workspace.ui.dto.response.CategoryUpdateResponse;
 import jakarta.validation.Valid;
-import java.net.URI;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -31,52 +23,20 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequiresRole(authRoles = {AuthRole.MEMBER})
-@RequestMapping("/api/workspaces/{workspaceIdentifier}/categories")
+@RequestMapping("/api/categories")
 public class CategoryController {
 
     private final CategoryService categoryService;
-
-    // 카테고리 생성
-    @PostMapping
-    public ResponseEntity<CategoryCreateResponse> createCategory(
-            final MemberAuthInfo memberAuthInfo,
-            @PathVariable final String workspaceIdentifier,
-            @Valid @RequestBody final CategoryCreateRequest request
-    ) {
-        final var response = categoryService.createCategory(
-                request.toCommandWith(memberAuthInfo, workspaceIdentifier)
-        );
-        return ResponseEntity.created(
-                URI.create("/api/workspaces/"
-                        + workspaceIdentifier + "/categories/"
-                        + response.id()
-                )
-        ).body(CategoryCreateResponse.from(response));
-    }
-
-    // 카테고리 순서 변경
-    @PostMapping("/sequence")
-    public ResponseEntity<CategoryReorderResponse> updateCategorySequence(
-            final MemberAuthInfo memberAuthInfo,
-            @PathVariable final String workspaceIdentifier,
-            @Valid @RequestBody final CategoryReorderRequest request
-    ) {
-        final var response = categoryService.updateCategorySequence(
-                request.toCommandWith(memberAuthInfo, workspaceIdentifier)
-        );
-        return ResponseEntity.ok(CategoryReorderResponse.from(response));
-    }
 
     // 카테고리 (이름/색상) 수정
     @PatchMapping("/{categoryId}")
     public ResponseEntity<CategoryUpdateResponse> updateCategory(
             final MemberAuthInfo memberAuthInfo,
-            @PathVariable final String workspaceIdentifier,
             @PathVariable final Long categoryId,
             @Valid @RequestBody final CategoryUpdateRequest request
     ) {
         final var response = categoryService.updateCategory(
-                request.toCommandWith(memberAuthInfo, workspaceIdentifier, categoryId)
+                request.toCommandWith(memberAuthInfo, categoryId)
         );
 
         return ResponseEntity.ok(CategoryUpdateResponse.from(response));
@@ -86,11 +46,10 @@ public class CategoryController {
     @DeleteMapping("/{categoryId}")
     public ResponseEntity<Void> deleteCategory(
             final MemberAuthInfo memberAuthInfo,
-            @PathVariable final String workspaceIdentifier,
             @PathVariable final Long categoryId
     ) {
         categoryService.deleteCategory(
-                new DeleteCategoryCommand(memberAuthInfo, workspaceIdentifier, categoryId)
+                new DeleteCategoryCommand(memberAuthInfo, categoryId)
         );
 
         return ResponseEntity.noContent().build();
@@ -100,26 +59,12 @@ public class CategoryController {
     @GetMapping("/{categoryId}")
     public ResponseEntity<CategoryResponse> readCategory(
             final MemberAuthInfo memberAuthInfo,
-            @PathVariable final String workspaceIdentifier,
             @PathVariable final Long categoryId
     ) {
         final var response = categoryService.findCategory(
-                new FindCategoryCommand(memberAuthInfo, workspaceIdentifier, categoryId)
+                new FindCategoryCommand(memberAuthInfo, categoryId)
         );
 
         return ResponseEntity.ok(CategoryResponse.from(response));
-    }
-
-    // 카테고리 목록 전체 조회 - 워크스페이스 상세 페이지
-    @GetMapping
-    public ResponseEntity<CategoriesResponse> readAllCategories(
-            final MemberAuthInfo memberAuthInfo,
-            @PathVariable final String workspaceIdentifier
-    ) {
-        final var response = categoryService.findAllCategories(
-                new FindAllCategoriesCommand(memberAuthInfo, workspaceIdentifier)
-        );
-
-        return ResponseEntity.ok(CategoriesResponse.from(response));
     }
 }
