+++ src/main/java/courseitda/workspace/ui/WorkspaceController.java
@@ -4,11 +4,18 @@ import courseitda.auth.domain.AuthRole;
 import courseitda.auth.domain.MemberAuthInfo;
 import courseitda.auth.domain.RequiresRole;
 import courseitda.member.domain.Member;
+import courseitda.workspace.application.CategoryService;
 import courseitda.workspace.application.WorkspaceService;
 import courseitda.workspace.application.dto.request.DeleteWorkspaceCommand;
+import courseitda.workspace.application.dto.request.FindAllCategoriesCommand;
 import courseitda.workspace.application.dto.request.ReadWorkspaceCommand;
+import courseitda.workspace.ui.dto.request.CategoryCreateRequest;
+import courseitda.workspace.ui.dto.request.CategoryReorderRequest;
 import courseitda.workspace.ui.dto.request.WorkspaceCreateRequest;
 import courseitda.workspace.ui.dto.request.WorkspaceUpdateRequest;
+import courseitda.workspace.ui.dto.response.CategoriesResponse;
+import courseitda.workspace.ui.dto.response.CategoryCreateResponse;
+import courseitda.workspace.ui.dto.response.CategoryReorderResponse;
 import courseitda.workspace.ui.dto.response.WorkspaceCreateResponse;
 import courseitda.workspace.ui.dto.response.WorkspaceReadResponse;
 import courseitda.workspace.ui.dto.response.WorkspaceUpdateResponse;
@@ -32,6 +39,51 @@ import org.springframework.web.bind.annotation.RestController;
 public class WorkspaceController {
 
     private final WorkspaceService workspaceService;
+    private final CategoryService categoryService;
+
+    // 카테고리 생성
+    @PostMapping("/{workspaceIdentifier}/categories")
+    public ResponseEntity<CategoryCreateResponse> createCategory(
+            final MemberAuthInfo memberAuthInfo,
+            @PathVariable final String workspaceIdentifier,
+            @Valid @RequestBody final CategoryCreateRequest request
+    ) {
+        final var response = categoryService.createCategory(
+                request.toCommandWith(memberAuthInfo, workspaceIdentifier)
+        );
+        return ResponseEntity.created(
+                URI.create("/api/workspaces/"
+                        + workspaceIdentifier + "/categories/"
+                        + response.id()
+                )
+        ).body(CategoryCreateResponse.from(response));
+    }
+
+    // 카테고리 순서 변경
+    @PostMapping("/{workspaceIdentifier}/categories/sequence")
+    public ResponseEntity<CategoryReorderResponse> updateCategorySequence(
+            final MemberAuthInfo memberAuthInfo,
+            @PathVariable final String workspaceIdentifier,
+            @Valid @RequestBody final CategoryReorderRequest request
+    ) {
+        final var response = categoryService.updateCategorySequence(
+                request.toCommandWith(memberAuthInfo, workspaceIdentifier)
+        );
+        return ResponseEntity.ok(CategoryReorderResponse.from(response));
+    }
+
+    // 카테고리 목록 전체 조회 - 워크스페이스 상세 페이지
+    @GetMapping("/{workspaceIdentifier}/categories")
+    public ResponseEntity<CategoriesResponse> readAllCategories(
+            final MemberAuthInfo memberAuthInfo,
+            @PathVariable final String workspaceIdentifier
+    ) {
+        final var response = categoryService.findAllCategories(
+                new FindAllCategoriesCommand(memberAuthInfo, workspaceIdentifier)
+        );
+
+        return ResponseEntity.ok(CategoriesResponse.from(response));
+    }
 
     // 워크스페이스 생성
     @PostMapping
